requirements.txt
# Cloud Account Manager - Requirements
PyQt6>=6.5.0
PyYAML>=6.0
pyotp>=2.8.0
boto3>=1.28.0
requests>=2.31.0
SQLAlchemy>=2.0.0
